RAG/requirements.txt
# Python Requirements for Ollama-RAG-Sync
# Install with: pip install -r requirements.txt

# PDF Processing Libraries
PyMuPDF>=1.23.0
marker-pdf>=0.2.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
ocrmypdf>=15.0.0

# Vector Database
chromadb>=0.4.0

# Core Dependencies
markdown>=3.4.0
